Rename to querybuilder.py
